Handle low disk space scenarios by resorting to unsafe write/rename
Since the safe operations take double the size of the target file as extra
space, if there is no enough space, they fail with `No space left on device
(os error 28)`.

In those situations when a copy or write operation fails with no space error,
we take a shortcut and resort to the unsafe counter parts of those operations.

Closes #120

Change-type: patch
Signed-off-by: Zahari Petkov <zahari@balena.io>