Add --unsafe/-u CLI argument
When unsafe fallback is enabled and operation fails because of out
of space error, fatrw will fallback to unsafe write/rename operations
that do not require extra space.

The default behavior will keep the safe behavior of fatrw, which means
that if an operation fails because of out of space error, fatrw will
not fallback to unsafe behavior and the operation will not succeed.

Change-type: patch
Signed-off-by: Zahari Petkov <zahari@balena.io>